Fix moved component references
